Phase 5 (2/N): split nest-format reader and writer into modules
reader.rs (413) → reader/{mod, parse, validate, decode}.rs
  mod.rs       72 — NestView struct + simple accessors (entry,
                    get_section_data, len, is_empty, raw_bytes)
  parse.rs    151 — from_bytes (header, section table, manifest,
                    footer parse + structural integrity)
  validate.rs 165 — semantic checks (required sections, embeddings
                    layout, search_contract cross-check, embeddings
                    NaN/Inf walk, encoding/section consistency)
  decode.rs    65 — decoded_section, search_contract, file_hash_hex,
                    content_hash_hex

writer.rs (626) → writer/{mod, encoding_choice, payload, build, tests}.rs
  mod.rs            134 — NestFileBuilder API + setters + write_to_path
  encoding_choice.rs 56 — SectionEncoding/EmbeddingDType + REPRODUCIBLE_CREATED
  payload.rs         55 — encode_embeddings_payload + maybe_zstd
  build.rs          213 — build_bytes orchestration
  tests.rs          218 — full builder + reader roundtrip tests

Largest format file is now manifest/mod.rs at 240 lines. All others
≤ 218. cargo test -p nest-format: 47 + 19 + 8 + many split-tests
green. Workspace 134 tests still pass.
