+++ crates/nest-format/src/writer/tests.rs
@@ -1,0 +1,218 @@
+//! Integration-style tests for the builder. Exercise the full
+//! build_bytes flow + reader roundtrip, plus the encoding / dtype
+//! invariants the runtime relies on.
+
+use super::*;
+use crate::chunk::ChunkInput;
+use crate::error::NestError;
+use crate::layout::{
+    SECTION_EMBEDDINGS, SECTION_ENCODING_FLOAT16, SECTION_ENCODING_INT8, SECTION_ENCODING_RAW,
+};
+use crate::manifest::Manifest;
+use crate::reader::NestView;
+
+fn good_manifest() -> Manifest {
+    Manifest {
+        embedding_model: "demo".into(),
+        embedding_dim: 4,
+        n_chunks: 1,
+        chunker_version: "demo-chunker/1".into(),
+        model_hash: format!("sha256:{}", "0".repeat(64)),
+        ..Default::default()
+    }
+}
+
+fn one_chunk() -> ChunkInput {
+    ChunkInput {
+        canonical_text: "hello world".into(),
+        source_uri: "doc.txt".into(),
+        byte_start: 0,
+        byte_end: 11,
+        embedding: vec![1.0, 0.0, 0.0, 0.0],
+    }
+}
+
+#[test]
+fn deterministic_build() {
+    let b1 = NestFileBuilder::new(good_manifest())
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    let b2 = NestFileBuilder::new(good_manifest())
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    assert_eq!(b1, b2);
+}
+
+#[test]
+fn reproducible_mode_overrides_created() {
+    let mut m1 = good_manifest();
+    m1.created = Some("2026-04-27T13:30:00Z".into());
+    let mut m2 = good_manifest();
+    m2.created = Some("2030-12-31T23:59:59Z".into());
+
+    let b1 = NestFileBuilder::new(m1)
+        .reproducible(true)
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    let b2 = NestFileBuilder::new(m2)
+        .reproducible(true)
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    assert_eq!(b1, b2, "reproducible builds must be byte-equal");
+
+    let view = NestView::from_bytes(&b1).unwrap();
+    assert_eq!(view.manifest.created.as_deref(), Some(REPRODUCIBLE_CREATED));
+}
+
+#[test]
+fn non_reproducible_preserves_created() {
+    let mut m = good_manifest();
+    m.created = Some("2026-04-27T13:30:00Z".into());
+    let bytes = NestFileBuilder::new(m.clone())
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    let view = NestView::from_bytes(&bytes).unwrap();
+    assert_eq!(
+        view.manifest.created.as_deref(),
+        Some("2026-04-27T13:30:00Z")
+    );
+}
+
+#[test]
+fn rejects_chunk_count_mismatch() {
+    let mut m = good_manifest();
+    m.n_chunks = 2;
+    let err = NestFileBuilder::new(m).add_chunk(one_chunk()).build_bytes();
+    assert!(matches!(err, Err(NestError::ManifestInvalid(_))));
+}
+
+#[test]
+fn rejects_bad_dim() {
+    let mut c = one_chunk();
+    c.embedding = vec![1.0, 0.0]; // wrong dim
+    let err = NestFileBuilder::new(good_manifest())
+        .add_chunk(c)
+        .build_bytes();
+    assert!(matches!(err, Err(NestError::DimensionMismatch { .. })));
+}
+
+#[test]
+fn zstd_encoding_preserves_content_hash() {
+    let mut m = good_manifest();
+    m.n_chunks = 3;
+    let chunks: Vec<ChunkInput> = (0..3)
+        .map(|i| ChunkInput {
+            canonical_text: format!("paragraph number {}", i),
+            source_uri: "doc.txt".into(),
+            byte_start: i * 10,
+            byte_end: i * 10 + 5,
+            embedding: {
+                let mut v = vec![0.0; 4];
+                v[(i % 4) as usize] = 1.0;
+                v
+            },
+        })
+        .collect();
+
+    let raw = NestFileBuilder::new(m.clone())
+        .add_chunks(chunks.clone())
+        .build_bytes()
+        .unwrap();
+    let zst = NestFileBuilder::new(m)
+        .text_encoding(SectionEncoding::Zstd)
+        .add_chunks(chunks)
+        .build_bytes()
+        .unwrap();
+
+    let v_raw = NestView::from_bytes(&raw).unwrap();
+    let v_zst = NestView::from_bytes(&zst).unwrap();
+    assert_eq!(
+        v_raw.content_hash_hex().unwrap(),
+        v_zst.content_hash_hex().unwrap(),
+        "content_hash must equal across wire encodings"
+    );
+    assert_ne!(
+        v_raw.file_hash_hex(),
+        v_zst.file_hash_hex(),
+        "file_hash must differ between raw and zstd"
+    );
+}
+
+#[test]
+fn float16_embeddings_roundtrip() {
+    let mut m = good_manifest();
+    m.n_chunks = 2;
+    let bytes = NestFileBuilder::new(m)
+        .embedding_dtype(EmbeddingDType::Float16)
+        .add_chunk(ChunkInput {
+            canonical_text: "a".into(),
+            source_uri: "d".into(),
+            byte_start: 0,
+            byte_end: 1,
+            embedding: vec![1.0, 0.0, 0.0, 0.0],
+        })
+        .add_chunk(ChunkInput {
+            canonical_text: "b".into(),
+            source_uri: "d".into(),
+            byte_start: 1,
+            byte_end: 2,
+            embedding: vec![0.0, 1.0, 0.0, 0.0],
+        })
+        .build_bytes()
+        .unwrap();
+    let view = NestView::from_bytes(&bytes).unwrap();
+    assert_eq!(view.manifest.dtype, "float16");
+    let entry = view.entry(SECTION_EMBEDDINGS).unwrap();
+    assert_eq!(entry.encoding, SECTION_ENCODING_FLOAT16);
+    assert_eq!(entry.size as usize, 2 * 4 * 2); // n*dim*2
+}
+
+#[test]
+fn int8_embeddings_roundtrip() {
+    let mut m = good_manifest();
+    m.n_chunks = 2;
+    let bytes = NestFileBuilder::new(m)
+        .embedding_dtype(EmbeddingDType::Int8)
+        .add_chunk(ChunkInput {
+            canonical_text: "a".into(),
+            source_uri: "d".into(),
+            byte_start: 0,
+            byte_end: 1,
+            embedding: vec![1.0, 0.0, 0.0, 0.0],
+        })
+        .add_chunk(ChunkInput {
+            canonical_text: "b".into(),
+            source_uri: "d".into(),
+            byte_start: 1,
+            byte_end: 2,
+            embedding: vec![0.0, 1.0, 0.0, 0.0],
+        })
+        .build_bytes()
+        .unwrap();
+    let view = NestView::from_bytes(&bytes).unwrap();
+    assert_eq!(view.manifest.dtype, "int8");
+    let entry = view.entry(SECTION_EMBEDDINGS).unwrap();
+    assert_eq!(entry.encoding, SECTION_ENCODING_INT8);
+    // 8-byte prefix + n*4 (scales) + n*dim
+    assert_eq!(entry.size as usize, 8 + 2 * 4 + 2 * 4);
+}
+
+#[test]
+fn rejects_zstd_embeddings() {
+    // Embeddings can never be zstd-compressed (we want SIMD-friendly
+    // mmap reads). text_encoding does not apply to embeddings.
+    let m = good_manifest();
+    let bytes = NestFileBuilder::new(m)
+        .text_encoding(SectionEncoding::Zstd)
+        .add_chunk(one_chunk())
+        .build_bytes()
+        .unwrap();
+    let view = NestView::from_bytes(&bytes).unwrap();
+    let entry = view.entry(SECTION_EMBEDDINGS).unwrap();
+    assert_eq!(entry.encoding, SECTION_ENCODING_RAW);
+}
